Almost finished.
* Could not set up the CSS switch;
* Table refresh is performing by separate button.
