--- j140-javafx-hw2/src/datamodel/CarException.java
@@ -1,17 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package datamodel;
-
-/**
- *
- * @author YuriPilshikov
- */
-public class CarException extends Exception{
-    public CarException(String message) {
-        super(message);
-    }
-    
-}
