--- j140-javafx-hw2/src/datamodel/UserException.java
@@ -1,16 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package datamodel;
-
-/**
- *
- * @author YuriPilshikov
- */
-public class UserException extends Exception{
-    public UserException(String message) {
-        super(message);
-    }
-}
