+++ j140-javafx-hw2/src/j140/javafx/fxml/datamodel/CarException.java
@@ -1,0 +1,17 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package j140.javafx.fxml.datamodel;
+
+/**
+ *
+ * @author YuriPilshikov
+ */
+public class CarException extends Exception{
+    public CarException(String message) {
+        super(message);
+    }
+    
+}
