+++ j140-javafx-hw2/src/j140/javafx/fxml/datamodel/User.java
@@ -1,0 +1,28 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package j140.javafx.fxml.datamodel;
+
+/**
+ *
+ * @author YuriPilshikov
+ */
+public class User {
+    private String name;
+    private String password;
+
+    public User(String name, String password) {
+        this.name = name;
+        this.password = password;
+    }    
+
+    public String getName() {
+        return name;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
