+++ j140-javafx-hw2/src/j140/javafx/fxml/datamodel/UserException.java
@@ -1,0 +1,16 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package j140.javafx.fxml.datamodel;
+
+/**
+ *
+ * @author YuriPilshikov
+ */
+public class UserException extends Exception{
+    public UserException(String message) {
+        super(message);
+    }
+}
