+++ j140-javafx-hw2/src/j140/javafx/fxml/resources/login.css
@@ -1,0 +1,21 @@
+/*
+ * Empty Stylesheet file.
+ */
+
+.label {
+    -fx-font-weight: bold;
+    -fx-padding: 15;
+}
+
+#bottomLabel {        
+        -fx-text-fill: red;
+        -fx-font-size: 20;
+        -fx-font-weight: bold;
+}
+
+.button {
+    -fx-font-weight: bold;
+}
+
+
+
